+++ corbeans-corda-webserver/src/test/kotlin/com/github/manosbatsis/corbeans/corda/webserver/InfoIntegrationTests.kt
@@ -26,7 +26,6 @@ import org.junit.jupiter.api.Test
 import org.slf4j.LoggerFactory
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.http.HttpStatus
-import org.springframework.web.client.RestTemplate
 
 
 open class InfoIntegrationTests(
@@ -53,7 +52,7 @@ open class InfoIntegrationTests(
         Assertions.assertEquals(HttpStatus.OK, swaggerUi.statusCode)
     }
 
-    //@Test
+    @Test
     fun `Can see Corda details within Actuator info endpoint response`() {
         logger.info("testInfoContributor, called")
         val entity = this.restTemplate
@@ -68,7 +67,7 @@ open class InfoIntegrationTests(
         validateCordaInfo(corda)
     }
 
-    //@Test
+    @Test
     fun `Can access Corda custom Actuator endpoint`() {
         logger.info("testCordaEndpoint, called")
         val serviceKeys = this.networkService.nodeServices.keys